+++ database/migrations/2025_04_23_111700_create_user_profiles_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('user_profiles', function (Blueprint $table) {
             $table->integer('profile_id', true);
-            $table->integer('user_id')->nullable()->index('user_id');
+            $table->unsignedBigInteger('user_id')->nullable()->index('user_id');
             $table->enum('pforce', ['1', '0'])->default('0')->index('pforce');
             $table->string('num_rows')->nullable()->index('num_rows');
             $table->string('num_filled')->nullable()->index('num_filled');
